added tests for DaoBaseAuto
